feat: 문제 데이터 패칭 함수
