+++ src/app/(auth)/(navigationsBarLayout)/problems/[problemId]/page.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@/components/ui/button';
 import { CodeEditor, ProblemSection, TerminalOutput } from '@/features/problem';
-import { getProblem } from '@/features/problem/actions/getProblem';
+import { getProblem } from '@/features/problem/actions/problem.actions';
 
 interface IProblemPageProps {
   params: Promise<{ problemId: string }>;
@@ -12,11 +12,13 @@ export default async function ProblemPage({ params }: IProblemPageProps) {
     'eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzIiwiZW1haWwiOiJnYnRteGxmQG5hdmVyLmNvbSIsInVzZXJuYW1lIjoi7Jyg7ISg7ZalIiwibmlja25hbWUiOiLjhYfjhYXjhY4iLCJ1c2VyUm9sZSI6IkFETUlOIiwidGllciI6Ik5FV0JJRSIsImV4cCI6MTc1MTg4MDIyMiwiaWF0IjoxNzUxMjc1NDIyfQ.Ud6bxBDvnGpOQeg48A86sBhvjqIUBvwGuL7HQJNII8s';
 
   const result = await getProblem(problemId, token);
+  // const testCase = await getTestCase(problemId, token);
 
   if (!result) {
     console.error('문제를 불러오는 데 실패했습니다.');
     return <div>문제를 불러오는 데 실패했습니다.</div>;
   }
+
   return (
     <main className="flex pt-20 h-full">
       <div className="flex-1 flex flex-col gap-[29px]">
@@ -30,7 +32,7 @@ export default async function ProblemPage({ params }: IProblemPageProps) {
       <section className="flex flex-col flex-1">
         <CodeEditor />
         <div className="h-[1px] w-full bg-white" />
-        <TerminalOutput />
+        <TerminalOutput problemId={problemId} token={token} />
       </section>
     </main>
   );
